Release v0.8.0

+++ src/generateSdk/utils/compileSdk.ts
@@ -37,11 +37,14 @@ export async function compileSdk(
     language: Language,
     publish: boolean,
     outDir = "../genezio-sdk",
+    overwriteIfExists = true,
 ) {
     const genezioSdkPath = path.resolve(sdkPath, outDir);
     // delete the old sdk
-    if (fs.existsSync(genezioSdkPath)) {
-        await deleteFolder(genezioSdkPath);
+    if (overwriteIfExists) {
+        if (fs.existsSync(genezioSdkPath)) {
+            await deleteFolder(genezioSdkPath);
+        }
     }
     fs.mkdirSync(genezioSdkPath, { recursive: true });
     // compile the sdk to cjs and esm using worker threads
